Fix wrong evalutaion of catalog constraints
Without the fix, a time constraint of the form "YM: 10" would be
accepted and interpreted as year 10 and minute 10 all at once.

This error was spotted while resolving mypy issues, see commit "Resolve
mypy issues in 'catalogue.py'" to come.

As an aside, make sure that GlobbingCreator._convert_to_relativedelta()
always returns a value of type 'relativedelta' or raises an exception,
but never returns None.