Reorder methods in Main to deal with two mypy issues
With this commit two methods of class Main are moved:
- Main._init_auto_restart() is now below Main.__init__()
- Main._restart_handler() is now far below Main._init_auto_restart()

The latter is important to get rid of
uwsift/__main__.py:566: error: Cannot determine type of "_restart_handler_active"  [has-type]
uwsift/__main__.py:571: error: Cannot determine type of "_last_restart_request"  [has-type]

Due to the way member variables of Main are defined in Main._init_*() methods
called via Main.__init__() the order of the two methods in question is
important for mypy to be able to determine the types of the two member
variables in question: they are used in Main._restart_handler(), but
declared/defined in Main._init_auto_restart(), thus mypy must see the
latter before the former.

It's not actually necessary to move Main._init_auto_restart() under
Main.__init__() to accomplish this, but it just makes more sense and
increases internal consistency.