Fix Algebraic layer loses color limits on repeated Update
For an algebraic layer, if the button "Update" in the "Algebraic" config
pane is clicked repeatedly without changing anything else, the layer
loses its color limits (displayd in the Layer Details pane for editing)
and its image in the scene vanishes, only to reappear when the button is
clicked again.

This happens because of an incorrect check if the layer has input
layers: It's color limits are set to INVALID_COLOR_LIMITS if *any* of its
up to 3 input layers is None, which is regularly the case e. g. for
formulas which take only 2 inputs. With the next click to "Update" the
then invalid color limits trigger a correct update again, which leads to
the described toggling.

Fix this, by only setting the color limits to INVALID_COLOR_LIMITS, if
the layer has no input layers at all (i.e., al input layer "slots" are
None).

TODO: This doesn't care for the case, that the configuraion of the
algebraic layer is invalid for another reason like n input layers
required by the formula but less than n selected - also in this case it
would be sensible to reset the color limits (if at all?!?).

Concerns #132