Enable to use data missing a valid range as RGB input
in this case, fall back to use the actual range of their layers for
setting the limits of the channel sliders in the "RGB Bounds" config
pane.

Until this fix (most likely, I haven't bothered to verify this
assessment) it was not possible to use data that has no valid range
defined as an input for recipe RGBs.

Related to #48