Fix OpenFileWizard Folder Selection fail on Windows
The customization of the file dialog as introduced with commit

0e35301d ("Enable tree navigation in "Select Data Directory" dialog", 2022-12-07)

can only work for Qt based file dialogs. But since Qt prefers to
instantiate the platform native file dialog via Qt's QFileDialog
interface this customization must only be applied,
if the actual file dialog provides a QTreeWidget for the files/directory
view.

The customiziation may only be applicable on KDE or other Qt based
Windows Managers on Linux, not on Windows and (not tested, but most
likely) not on MacOS. Also on GTK based Windows Managers (e.g. Gnome) it
may not be available.

Fixes #161