Fix timeline click fails for irregular time steps
Revise the code associating a time step mark with the current mouse
position at click. The idea behind the implementation as read from the
code is elegant, but the actual implementation is just broken:
calculated distances were not actually used to find the mark closest to
the mouse position and ... whatever.

Fixes #160