Beautify timeline ruler
Make the ruler look sleeker and avoid antialiasing artefacts for lines
and text.

This is achieved by fine-tuning the drawing positions and the line
widths: Set the line width of ruler and tick marks to 1 pixel and make
sure that they are actually drawn on *one* pixel by adding an offset of
0.5.  This is necessary because the coordinate of a pixel defines its
upper left corner and a line of width 1 is centered around this
coordinate if no offset is given. Such a line would touch two adjacent
pixels which with antialing are rendered in halftone.