Enable to fit the main window on Full HD screens
With the integration of the Layer Details (#132) and the Dataset
Statistics (#59) panes the layout takes a little too much vertical
space.

Thus decrease the minimal height of the Layer Manager pane which in
consequence allows to resize the main window to a height less than 1080
pixels.

Unfortunately the initial size seems to be still larger.

TODO: Find a way (using QDesigner?) to restrict the Main Window height to
less than screen height.