Improve initial fit of MainWindow to the screen
The revised function set_default_geometry() tries to resize the main
window so that it fits on the screen (which may not be possible, if the
screen is small, Full HD resolution is necessary at least) and center it
as good as possible.

See the TODO comment in the implementation: since some of the parameters
required to do the fit are not available until after the Window Manager
of the OS has finished setting up the top level widget, which happens
asynchronously, the current implementation cannot do the best it can.

The implementation is completely revised and avoids deprecated Qt API.
It also makes the --desktop command line argument work again.