Prepare for Workspace classes interface cleanup
Only CachingWorkspace needs the caching related interface. Several
client methods are (and must) only be called, when the instantiated
Workspace is a CachingWorkspace (which is the case, with the
configuration setting "storage.use_inventory_db: True").  Make this
explicit by assertions which check the type of the Workspace.

This secures it to remove the methods which are specific to the caching
use case from BaseWorkspace and SimpleWorkspace, which is foreseen for
the next commit.