Clean up {Base,Caching}Workspace implementation
Remove class variables that will later be overshadowed by member
variables of the same name anyway when the latter are initialized - the
old code from early SIFT development is obviously the result of a
misconception about class and member variables in Python.

Clean up the initialization of BaseWorkspace._queue: the variable is
used in BaseWorkspace methods, but only initialized in the derived class
CachingWorkspace. This is improved by passing through the initialization
given to CachingWorkspace.__init__() to BaseWorkspace.__init__().

The matter was spotted during mypy issue resolution but it belongs
better to #89 in terms of its concern.