Fix stuck point probe text in status line
To achieve this, revise the updating of the point probe text.  There is
a problem when selecting a layer that has no point probe value, e.g.
"MC_IMAGE" ([M]ulti [C]hannel Image).
This problem was introduced with 28dbb863
("Show the point probe information of the selected layer", 2022-09-26).

Make it possible to select any type of layer again without errors when
the point probe is active/set.

Fixes #163

Additionally:
- When a probe is active always display the name of the selected layer
  in the status bar, even if everything else evalutates to N/A
- Implementation detail: make the implemetation of the processing of the
  point probe value to the unit converted and formatted display string
  more verbose in order to make it more comprehensible.