Enable calculation of additional statistical metrics for difference algebraics
The statistics module does not calculate the additional metrics for the
difference algebraic, because the comparison of the formula term with
'x-y' does not succeed for the preset "Difference" formula 'x - y'.

Removing all spaces from the term of the operation formula fixes this
problem.

Also enable the calculation of the additional statistical metrics for
the commuted term 'y-x'.

Fixes #179

Furthermore remove unneeded update of the info dict before calculating
statistics.