Improve statistics table layout
When the content of the statistics table is updated, resize all but the
last column to the width needed to show its content completely. The last
column is stretched to fill the parent widget horizontally.

The column widths can be changed manually, but this configuration is
lost with the next content update.

Fixes #177