Remove that ints are padded with spaces in statistics table
As many trailing spaces are added to integer values as decimal places
are configured to display for floats, e.g.

|  1     |
|  1.234 |

This looked like good idea to make the magnitudes of ints and floats
easier to compare, but this leads to a waste of space if there are no
floats but only ints to display in a column, which is usually the case
for categorial data.

Thus: do not pad ints according to the configured number of decimal
places any more.