Get rid of duplicate colormaps list entry 'grays'
Currently the colormap 'grays' is added to the OTHER_COLORMAPS but is
also contained in the 'useable_vispy_cmaps' which leads to the
duplicate.

Two entries of the same name lead to unexpected behaviour of the
according combo box: Whenever the second of the identical entries is
selected actually the first one becomes selected.

Fix this by removing 'grays' from the OTHER_COLORMAPS list.

This bug was introduced with commit
454dfad7 ("Update colormap handling to use lazy file loading and site/user dirs", 2018-06-27)
when VisPy colortables were added as group without removing the then
duplicate explicit entry of 'grays'.

Fixes #170