Introduce simple (non-tiling) geolocated image display
... as third way to display image data (i.e., data on a regular
rectangular grid with an Satpy AreaDefinition), which works as follows:

The raster data is put as texture on a subdivided quadrangle whose
vertices are transformed by the projection transformation in the
scenegraph (the required functionality is provided by the VisPy
scnenegraph node 'Image' and its variants used in SIFT).  The
subdivision resolution is hard-coded in this commit and will be made
configurable in future commits, as it can be used to balance quality
against resource consumption.

To be able to choose between the two old and the new image display mode,
the way the modes are controlled needs to be revised. 2 modes could be
distinguished with a boolean, but for 3 modes let's introduce 3 mutually
exclusive(*) options:

- "simple_geolocated" - the new mode
- "tiled_geolocated" - the revised name (now config value(!))for the
  former default mode
- "pixel_matrix" - the newly introduced name for the former
  non-"tiled_geolocated" image display

Represent these options in a Enum ImageDisplayMode and replace checks
against USE_TILED_GEOLOCATED_IMAGES with checks against the new
"constant" IMAGE_DISPLAY_MODE.

For details regarding the configuration to select one of these modes see
the updates to the documentation which will come in future commits.

(*) With appropriate user controls in place modes could be switched
interactively or even different data could be presented in different
modes, but this is out of scope for now.