Fix conda setup without xRITDecompress fails under Windows
When you try to create a Conda environment without xRITDecompress (which
is the default if you use Conda and not Mamba) on Windows by running the
according build target 'devel-bootstrap' you get the error

 process_begin: CreateProcess(NULL, true install --name devel-3.10 --channel eumetsat --strict-channel-priority public-decomp-wt, ...) failed.

This is because on Windows there is no command `true` (which on Linux
always succeeds while ignoring all command line arguments). Thus the
trick used to selectively install or not install the package
'public-decomp-wt' (by either runnig `conda install ...` or `true
install ...` depending on the setting of WITH_XRITDecompress) in commit

548de41c ("Integrate xRITDecompress into Conda environment and application", 2022-11-21)

breaks on Windows.

As a solution take advantage of CMake's platform abstraction support
with `cmake -E {env,true}` (see
https://cmake.org/cmake/help/latest/manual/cmake.1.html#run-a-command-line-tool).

Concerns #138