Reduce memory leakage after removing layers/datasets
One approach that turned out to lead to increased main memory freeing
after a layer (and as a consequence its datasets) has been removed is to
delete the corresponding memmap files. This is done when the setting
"storage.cleanup_file_cache" is configured to "True" (only available
when "storage.use_inventory_db: False", i.e., when the SimpleWorkspace
is in use; the CachingWorkspace relies on the persistence of the memmap
files).

This approach has an issue on Windows though: File system locks prevent
the files to be removed at most places in the code where it was done
leading to PermissionError exceptions.

Therefore, revise execution of file cache clean-up:
- Catch and ignore this error when it happens (just a DEBUG level log
  message is given for now).
- Try again to remove cached files after the garbage collector has run,
  which is explicity triggered after a dataset deletion, has finished.
  Here the file locks should eventually be released.

Should bring about an improvement in #182