Use 'nominal_start_time' as SCHED_TIME of dataset
... if the Satpy reader provides this attribute in the dataset's
metadata. Otheriwse fall back to the value of 'start_time' which is used
only before this commit.

The SCHED_TIME is used by the current TimeManager system to define the
order of datasets on the timeline and define the start of their "active
period" (i.e. the span of a dataset on the timeline in which it is
displayed).

Contributes to #185 in the following sense:

With this commit in place, it is sufficient to improve Satpy readers to
provide a "correct" nominal_start_time to allow the creation of (some)
algebraics between MSG SEVIRI and MTG FCI data: Their actual observation
times are not exactly on full (5,10,15) minutes, but may be a few
seconds behind the nominal start time, depending on the satellite.
Thus, as the current time matching only considers datasets with exactly
the same SCHED_TIME to be combinable, using the nominal start time
instead of the (observation?) start time makes it possible to combine
e.g. MSG and MTG FullDisk at least every 30 minutes into an Algebraic.

TODO:
Revise the mapping between Satpy times (nominal_start,end_time,
observation_start,end_time, start,end_time) and SIFT times (SCHED_TIME,
OBS_TIME, OBS_DURATION).