Make simple geolocated image grid resolution configurable
and somewhat adaptive: the size of the grid cells is calculated based on
the wanted cell size configured in m with the settings
'display.grid_cell_width' and 'display.grid_cell_height' similar to the
pixel size in area definitions.  The number of grid cells in x and y
direction is then calculated to approximately yield cells of the
configured size (again in the same sense as the pixel size in area
definitions).