For Algebraics, fit color map to actual range
... instead of to the VALID_RANGE metadata of the layer.

In certain cases, the calculation of the valid range for an algebraic
based on the valid ranges of the input data appears to yield much larger
ranges than the actual data spans. This leads to inappropriate colour
mapping: no features become visible until the colour limits are manually
adjusted (e.g. via Fit to Data).

So: ignore the VALID_RANGE of the Algebraic data/layer, but use the
min/max value pair calculated over all data of the layer when initally
determining the colour limits.

Fixes #187