Open first "Area Probe Graph" tab only on request
... i.e. when either the "Region" button or its menu item "Select Full
Data" is clicked.

The reason for this is that the Matplotlib toolbar should not be visible
when the application is started. It has been reported that users
misinterpret the pan button (the arrow cross) as the button for
selecting the pan mode in the main scene. This can be avoided by not
displaying an empty probe graph initially.

Closes #95