Automatically select new layer in LayerManager
Whenever a new layer is added, either by loading new data or creating a
composite layer, it is automatically selected in the LayerManager.

As a result, the dataset statistics of the new layer are automatically
displayed without any further user interaction, for example. NOTE: The
latter is not the case if the loaded dataset does not become 'active'
(i.e. not displayed) due to the current timeline settings (driving layer
and display time).