Allow users to close pane DockWidgets
A disadvantage of this option: Users may close a pane and not know how
to get it back, although it is easy: right click on the application menu
bar shows an according context menu.

If you want to prevent users from closing selected panes, such as the
"Layer Manager" or "Statistics" panes, this commit shows the right place
to start: just create a second list of panes and call setFeatures() on
them accordingly.

Contributes to #175