Fix failure to init Algebraics with mixed input sources
... i.e., when the data comes from different satellites or instruments.

ABI_AHI_Guidebook._default_display_time() raises KeyError exceptions if
PLATFORM or INSTRUMENT are not set.

On the other hand, these are not set in
BaseWorkspace._get_composite_metadata(), if they are not the same for
all inputs (to an algebraic).

Solve this by introducing MIXED = "Mixed" as pseudo PLATFORM and pseudo
INSTRUMENT and use them to set the according entries for mixed-source
algebraics metadata.

Fixes #193