Fix Simple Geolocated image display for data resampled to Plate Carree
When "Plate Carree" is the target projection for resampling during
loading the resulting dataset cell size information
`info[Info.CELL_{WIDTH,SIZE}]` is not in metres but degrees.

In SceneGraphManager._calc_subdivision_grid() this is not taken into
account, leading to huge numbers of pixels_per_cell_* and inversely to
num_cells_* = 0 and in the next step to a DivisionByZero exception.

Fix this by implementing a unit conversion from degrees to metres for
this special case.