Revise ABI_AHI_Guidebook._default_display_name(), again
Remove some reuse of variables and name aliasing (e.g. rename "sat" ->
"platform" as it takes the value info[Info.PLATFORM]).

In doing so, I believe I have understood the original intent regarding
the mapping of "toa_bidirectional_reflectance" and
"toa_brightness_temperature" to "Refl" and "BT": They are added to the
short_name in abbreviated form only for the two given standard_names,
otherwise there should not be such a label. So the previous commit,
3e8d86db ("Fix ABI_AHI_Guidebook._default_display_time()'s interpretaion of standard_name", 2023-02-09)
did no good.

Change the code to re-implement the original intent.