Fix that OpenFileWizard ignores path typed into DirectoryDialog
Work around a QFileDialog flaw that typing in the "Directory" edit field
highlights matching folders but does not seem to send the according
signal QFileDialog.directoryEntered or QFileDialog.currentChanged as you
would expect.

At least when hitting [Return] to confirm the entered path - which
unfortunately closes the QFileDialog since it is activating the "Choose"
button - the path should be sent to the OpenFileWizard to let it update
the "Folder" entry and the list of files in the folder matching the
current "Filter" pattern.

Unfortunately, since no signals seem to be sent while the user enters
the path in the "Directory" edit field, the OpenFileWizard cannot be
noticed to update its "Folder" and matching files list on the fly as it
happens when navigation through the directory tree with the mouse.