Start OpenFileWizard's DirectoryDialog from last chosen folder
When the DirectoryDialog from the OpenFileWizard is opened again after a
path has been entered in the "Directory" field, it starts again from
where it started before - not from the entered path! - unlike when the
path was navigated to by mouse interaction.

Work around this by explicitly setting the start directory of the dialog
to the navigated path.

The disadvantage of this is that the Directory Dialog will not open in
exactly the same state as when it was left, i.e. the root of the tree
shown in the central directory list is the navigated path, not the
ancestor folder as when the dialogue was last closed.