Fixed valid min/max handing when valid_range is None (fixed by Ana)
For valid min and valid max, the minimum and maximum values are taken from actual and valid range respectively. However, in the case when the valid range is None, this leads to a bug: valid_min is greater than valid_max.
Fixed this with this in the exception handler of range_hull_no_fail.